refactor: 디렉토리 변경 mvc 패턴적용
