+++ server.js
@@ -1,14 +1,14 @@
 const express = require('express')
 const app = express()
 const nunjucks = require('nunjucks')
-const router = require('./routes')
+const router = require('./src/routes')
 
 app.use(express.urlencoded({extended:true}))
 app.use(express.json())
-app.use(express.static('public'))
+app.use(express.static('./src/public'))
 // app.use(express.static('public/j'))
 app.set('view engine', 'html')
-nunjucks.configure('views',{
+nunjucks.configure('./src/views',{
   express:app,
   watch:true,
 })
