+++ src/routes/auth/index.js
