+++ src/routes/readme.md
@@ -1,0 +1,3 @@
+## 컨트롤러는 각 라우터의 디렉토리와 같은 위치에 있음.
+
+## main : 첫 페이지
