+++ src/views/admin/board.html
