+++ src/views/board/list.html
