+++ src/views/board/view.html
