+++ src/views/board/write.html
